+++ examples/dashboard/java/PanelLegend.java
@@ -15,7 +15,8 @@ public class PanelLegend extends Panel {
         shortcuts.put("N", "New Window");
         shortcuts.put("T", "Toggle Titlebar");
         shortcuts.put("W", "Close Window");
-        shortcuts.put("F", "Clipboard formats");
+        shortcuts.put("F", "Fullscreen");
+        shortcuts.put("X", "Clipboard formats");
         shortcuts.put("Z", "Toggle Z-order");
         shortcuts.put("B", "Toggle Progress Bar");
         shortcuts.put("O", "Opacity");