+++ shared/cc/impl/Library.cc
@@ -404,6 +404,30 @@ namespace jwm {
             }
         }
 
+        namespace EventWindowFullScreenEnter {
+            jobject kInstance;
+
+            void onLoad(JNIEnv* env) {
+                JNILocal<jclass> cls(env, env->FindClass("io/github/humbleui/jwm/EventWindowFullScreenEnter"));
+                Throwable::exceptionThrown(env);
+                jfieldID field = env->GetStaticFieldID(cls.get(), "INSTANCE", "Lio/github/humbleui/jwm/EventWindowFullScreenEnter;");
+                jobject instance = env->GetStaticObjectField(cls.get(), field);
+                kInstance = env->NewGlobalRef(instance);
+            }
+        }
+
+        namespace EventWindowFullScreenExit {
+            jobject kInstance;
+
+            void onLoad(JNIEnv* env) {
+                JNILocal<jclass> cls(env, env->FindClass("io/github/humbleui/jwm/EventWindowFullScreenExit"));
+                Throwable::exceptionThrown(env);
+                jfieldID field = env->GetStaticFieldID(cls.get(), "INSTANCE", "Lio/github/humbleui/jwm/EventWindowFullScreenExit;");
+                jobject instance = env->GetStaticObjectField(cls.get(), field);
+                kInstance = env->NewGlobalRef(instance);
+            }
+        }
+
         namespace EventWindowMove {
             jclass kCls;
             jmethodID kCtor;
@@ -571,6 +595,8 @@ extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_jwm_impl_Library__1nAf
     jwm::classes::EventWindowMinimize::onLoad(env);
     jwm::classes::EventWindowFocusIn::onLoad(env);
     jwm::classes::EventWindowFocusOut::onLoad(env);
+    jwm::classes::EventWindowFullScreenEnter::onLoad(env);
+    jwm::classes::EventWindowFullScreenExit::onLoad(env);
     jwm::classes::EventWindowMove::onLoad(env);
     jwm::classes::EventWindowResize::onLoad(env);
     jwm::classes::EventWindowRestore::onLoad(env);