+++ shared/java/EventWindowFullScreenEnter.java
@@ -1,0 +1,8 @@
+package io.github.humbleui.jwm;
+
+import lombok.Data;
+
+@Data
+public class EventWindowFullScreenEnter implements Event {
+    public static final EventWindowFullScreenEnter INSTANCE = new EventWindowFullScreenEnter();
+}