+++ shared/java/EventWindowFullScreenExit.java
@@ -1,0 +1,8 @@
+package io.github.humbleui.jwm;
+
+import lombok.Data;
+
+@Data
+public class EventWindowFullScreenExit implements Event {
+    public static final EventWindowFullScreenExit INSTANCE = new EventWindowFullScreenExit();
+}